streamlit/requirements.txt
streamlit
requests
pillow
pdf2image
